+++ tests/Mock/GsonMockResponse.php
@@ -1,0 +1,20 @@
+<?php
+/*
+ * Copyright (c) Nate Brunette.
+ * Distributed under the MIT License (http://opensource.org/licenses/MIT)
+ */
+
+namespace Tebru\Gson\Test\Mock;
+
+use Tebru\Gson\Annotation as Gson;
+
+/**
+ * Class GsonMockResponse
+ *
+ * @Gson\VirtualProperty("data")
+ *
+ * @author Nate Brunette <n@tebru.net>
+ */
+class GsonMockResponse extends GsonMock
+{
+}