+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 
 This document keeps track of changes between releases of the library.
 
+master
+------
+
+* Allowing type override during serialization
+
 v0.6.0
 ------